Changed the new spell blueprints so that they are always added to the game regardless if the feat is toggled on or off.
